// packages/fe-job-pc/src/pages/admin/service/hierarchical-diagnosis/diagnosis/interface.d.ts
export interface MeasureItem {
    code?: string
    title?: string
    type?: number
    enableState?: number,
    level?: number[],
    serviceNum?: number
}